Replace structs by-value usages with by-ref usages (#95)
Struct by-value usage incurs the cost of copying/cloning the struct data which should be avoided where it is not required.

P.S.
The CreateWorkflow function is broken. It was supposed to assign an increasing id to the primary and on-failure steps. But the existing implementation didn't work as expected. So, now I added a test case to verify the expected behaviour and also fixed the CreateWorkflow function.